multitenant-auth/requirements.txt
# requirements.txt
Flask
flask-cors
PyJWT
pymongo
redis
python-dotenv
requests
python-jose[cryptography]
